checkout of user's pet feature has been implemented

+++ src/actions/actions.ts
@@ -102,6 +102,13 @@ export async function editPet(petId: unknown, petData: unknown) {
 export async function deletePet(petId: unknown) {
   await sleep(1000);
 
+  // authentication check
+  const session = await auth();
+  if (!session?.user) {
+    redirect("/login");
+  }
+
+  // validation
   const validatedPetId = petIdSchema.safeParse(petId);
 
   if (!validatedPetId.success) {
@@ -110,6 +117,22 @@ export async function deletePet(petId: unknown) {
     };
   }
 
+  // authorization check
+  const pet = await prisma.pet.findUnique({
+    where: {
+      id: validatedPetId.data,
+    },
+  });
+  if (!pet) {
+    return { message: "Pet not found." };
+  }
+  if (pet.userId !== session.user.id) {
+    return {
+      message: "Not authorized.",
+    };
+  }
+
+  // database manipulation
   try {
     await prisma.pet.delete({
       where: {
